# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HelpFinance API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
} as const;

export type TransactionRecurringInterval =
  | (typeof TransactionRecurringInterval)[keyof typeof TransactionRecurringInterval]
  | null;

export const TransactionRecurringInterval = {
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
  yearly: "yearly",
} as const;

export interface Transaction {
  id: number;
  description: string;
  amount: number;
  type: TransactionType;
  category: string;
  date: string;
  isRecurring: boolean;
  recurringInterval?: TransactionRecurringInterval;
  createdAt: string;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  income: "income",
  expense: "expense",
} as const;

export type CreateTransactionBodyRecurringInterval =
  | (typeof CreateTransactionBodyRecurringInterval)[keyof typeof CreateTransactionBodyRecurringInterval]
  | null;

export const CreateTransactionBodyRecurringInterval = {
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
  yearly: "yearly",
} as const;

export interface CreateTransactionBody {
  description: string;
  amount: number;
  type: CreateTransactionBodyType;
  category: string;
  date: string;
  isRecurring: boolean;
  recurringInterval?: CreateTransactionBodyRecurringInterval;
}

export type CategoryType = (typeof CategoryType)[keyof typeof CategoryType];

export const CategoryType = {
  income: "income",
  expense: "expense",
  both: "both",
} as const;

export interface Category {
  id: number;
  name: string;
  icon: string;
  color: string;
  type: CategoryType;
}

export interface DashboardSummary {
  totalBalance: number;
  monthlyIncome: number;
  monthlyExpenses: number;
  savingsRate: number;
  topExpenseCategory: string;
  transactionCount: number;
  aiInsight: string;
}

export interface MonthlyTrendItem {
  month: string;
  income: number;
  expenses: number;
  savings: number;
}

export interface CategoryBreakdownItem {
  category: string;
  amount: number;
  percentage: number;
  color: string;
}

export interface Badge {
  id: string;
  name: string;
  description: string;
  icon: string;
  earned: boolean;
  earnedAt?: string | null;
  xpReward: number;
}

export interface GamificationProfile {
  xp: number;
  level: number;
  levelName: string;
  xpToNextLevel: number;
  streak: number;
  badgeCount: number;
  earnedBadges: Badge[];
}

export interface Goal {
  id: number;
  title: string;
  targetAmount: number;
  currentAmount: number;
  deadline?: string | null;
  category: string;
  icon: string;
  color: string;
  createdAt: string;
}

export interface CreateGoalBody {
  title: string;
  targetAmount: number;
  currentAmount: number;
  deadline?: string | null;
  category: string;
  icon: string;
  color: string;
}

export interface UpdateGoalBody {
  currentAmount?: number;
  title?: string;
  targetAmount?: number;
}

export interface AiChatBody {
  message: string;
  context?: string | null;
}

export interface AiChatResponse {
  reply: string;
  suggestions: string[];
}

export type ListTransactionsParams = {
  month?: string;
  category?: string;
  type?: ListTransactionsType;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  income: "income",
  expense: "expense",
} as const;

export type GetDashboardSummaryParams = {
  month?: string;
};

export type GetCategoryBreakdownParams = {
  month?: string;
};