# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HelpFinance API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: transactions
    description: Transaction operations
  - name: categories
    description: Category operations
  - name: dashboard
    description: Dashboard summary operations
  - name: gamification
    description: Gamification operations
  - name: goals
    description: Financial goals
  - name: ai
    description: AI assistant chat
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary
      parameters:
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/monthly-trend:
    get:
      operationId: getMonthlyTrend
      tags: [dashboard]
      summary: Get monthly income vs expense trend (last 6 months)
      responses:
        "200":
          description: Monthly trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyTrendItem"

  /dashboard/category-breakdown:
    get:
      operationId: getCategoryBreakdown
      tags: [dashboard]
      summary: Get expense breakdown by category
      parameters:
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdownItem"

  /gamification/profile:
    get:
      operationId: getGamificationProfile
      tags: [gamification]
      summary: Get user gamification profile (XP, level, badges, streak)
      responses:
        "200":
          description: Gamification profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GamificationProfile"

  /gamification/badges:
    get:
      operationId: listBadges
      tags: [gamification]
      summary: List all available badges
      responses:
        "200":
          description: List of badges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Badge"

  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List financial goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a financial goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"

  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update goal progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"

  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Send a message to AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Transaction:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string
        isRecurring:
          type: boolean
        recurringInterval:
          type: string
          nullable: true
          enum: [daily, weekly, monthly, yearly, null]
        createdAt:
          type: string
      required:
        - id
        - description
        - amount
        - type
        - category
        - date
        - isRecurring
        - createdAt

    CreateTransactionBody:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string
        isRecurring:
          type: boolean
        recurringInterval:
          type: string
          nullable: true
          enum: [daily, weekly, monthly, yearly, null]
      required:
        - description
        - amount
        - type
        - category
        - date
        - isRecurring

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [income, expense, both]
      required:
        - id
        - name
        - icon
        - color
        - type

    DashboardSummary:
      type: object
      properties:
        totalBalance:
          type: number
        monthlyIncome:
          type: number
        monthlyExpenses:
          type: number
        savingsRate:
          type: number
        topExpenseCategory:
          type: string
        transactionCount:
          type: integer
        aiInsight:
          type: string
      required:
        - totalBalance
        - monthlyIncome
        - monthlyExpenses
        - savingsRate
        - topExpenseCategory
        - transactionCount
        - aiInsight

    MonthlyTrendItem:
      type: object
      properties:
        month:
          type: string
        income:
          type: number
        expenses:
          type: number
        savings:
          type: number
      required:
        - month
        - income
        - expenses
        - savings

    CategoryBreakdownItem:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        percentage:
          type: number
        color:
          type: string
      required:
        - category
        - amount
        - percentage
        - color

    GamificationProfile:
      type: object
      properties:
        xp:
          type: integer
        level:
          type: integer
        levelName:
          type: string
        xpToNextLevel:
          type: integer
        streak:
          type: integer
        badgeCount:
          type: integer
        earnedBadges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
      required:
        - xp
        - level
        - levelName
        - xpToNextLevel
        - streak
        - badgeCount
        - earnedBadges

    Badge:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        earned:
          type: boolean
        earnedAt:
          type: string
          nullable: true
        xpReward:
          type: integer
      required:
        - id
        - name
        - description
        - icon
        - earned
        - xpReward

    Goal:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: string
          nullable: true
        category:
          type: string
        icon:
          type: string
        color:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - targetAmount
        - currentAmount
        - category
        - icon
        - color
        - createdAt

    CreateGoalBody:
      type: object
      properties:
        title:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: string
          nullable: true
        category:
          type: string
        icon:
          type: string
        color:
          type: string
      required:
        - title
        - targetAmount
        - currentAmount
        - category
        - icon
        - color

    UpdateGoalBody:
      type: object
      properties:
        currentAmount:
          type: number
        title:
          type: string
        targetAmount:
          type: number
      required: []

    AiChatBody:
      type: object
      properties:
        message:
          type: string
        context:
          type: string
          nullable: true
      required:
        - message

    AiChatResponse:
      type: object
      properties:
        reply:
          type: string
        suggestions:
          type: array
          items:
            type: string
      required:
        - reply
        - suggestions