# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HelpFinance API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all transactions
 */
export const ListTransactionsQueryParams = zod.object({
  month: zod.coerce.string().optional(),
  category: zod.coerce.string().optional(),
  type: zod.enum(["income", "expense"]).optional(),
});

export const ListTransactionsResponseItem = zod.object({
  id: zod.number(),
  description: zod.string(),
  amount: zod.number(),
  type: zod.enum(["income", "expense"]),
  category: zod.string(),
  date: zod.string(),
  isRecurring: zod.boolean(),
  recurringInterval: zod
    .union([
      zod.literal("daily"),
      zod.literal("weekly"),
      zod.literal("monthly"),
      zod.literal("yearly"),
      zod.literal(null),
    ])
    .nullish(),
  createdAt: zod.string(),
});
export const ListTransactionsResponse = zod.array(ListTransactionsResponseItem);

/**
 * @summary Create a transaction
 */
export const CreateTransactionBody = zod.object({
  description: zod.string(),
  amount: zod.number(),
  type: zod.enum(["income", "expense"]),
  category: zod.string(),
  date: zod.string(),
  isRecurring: zod.boolean(),
  recurringInterval: zod
    .union([
      zod.literal("daily"),
      zod.literal("weekly"),
      zod.literal("monthly"),
      zod.literal("yearly"),
      zod.literal(null),
    ])
    .nullish(),
});

/**
 * @summary Get a transaction
 */
export const GetTransactionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTransactionResponse = zod.object({
  id: zod.number(),
  description: zod.string(),
  amount: zod.number(),
  type: zod.enum(["income", "expense"]),
  category: zod.string(),
  date: zod.string(),
  isRecurring: zod.boolean(),
  recurringInterval: zod
    .union([
      zod.literal("daily"),
      zod.literal("weekly"),
      zod.literal("monthly"),
      zod.literal("yearly"),
      zod.literal(null),
    ])
    .nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a transaction
 */
export const DeleteTransactionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  icon: zod.string(),
  color: zod.string(),
  type: zod.enum(["income", "expense", "both"]),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary Get dashboard summary
 */
export const GetDashboardSummaryQueryParams = zod.object({
  month: zod.coerce.string().optional(),
});

export const GetDashboardSummaryResponse = zod.object({
  totalBalance: zod.number(),
  monthlyIncome: zod.number(),
  monthlyExpenses: zod.number(),
  savingsRate: zod.number(),
  topExpenseCategory: zod.string(),
  transactionCount: zod.number(),
  aiInsight: zod.string(),
});

/**
 * @summary Get monthly income vs expense trend (last 6 months)
 */
export const GetMonthlyTrendResponseItem = zod.object({
  month: zod.string(),
  income: zod.number(),
  expenses: zod.number(),
  savings: zod.number(),
});
export const GetMonthlyTrendResponse = zod.array(GetMonthlyTrendResponseItem);

/**
 * @summary Get expense breakdown by category
 */
export const GetCategoryBreakdownQueryParams = zod.object({
  month: zod.coerce.string().optional(),
});

export const GetCategoryBreakdownResponseItem = zod.object({
  category: zod.string(),
  amount: zod.number(),
  percentage: zod.number(),
  color: zod.string(),
});
export const GetCategoryBreakdownResponse = zod.array(
  GetCategoryBreakdownResponseItem,
);

/**
 * @summary Get user gamification profile (XP, level, badges, streak)
 */
export const GetGamificationProfileResponse = zod.object({
  xp: zod.number(),
  level: zod.number(),
  levelName: zod.string(),
  xpToNextLevel: zod.number(),
  streak: zod.number(),
  badgeCount: zod.number(),
  earnedBadges: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      description: zod.string(),
      icon: zod.string(),
      earned: zod.boolean(),
      earnedAt: zod.string().nullish(),
      xpReward: zod.number(),
    }),
  ),
});

/**
 * @summary List all available badges
 */
export const ListBadgesResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string(),
  icon: zod.string(),
  earned: zod.boolean(),
  earnedAt: zod.string().nullish(),
  xpReward: zod.number(),
});
export const ListBadgesResponse = zod.array(ListBadgesResponseItem);

/**
 * @summary List financial goals
 */
export const ListGoalsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  targetAmount: zod.number(),
  currentAmount: zod.number(),
  deadline: zod.string().nullish(),
  category: zod.string(),
  icon: zod.string(),
  color: zod.string(),
  createdAt: zod.string(),
});
export const ListGoalsResponse = zod.array(ListGoalsResponseItem);

/**
 * @summary Create a financial goal
 */
export const CreateGoalBody = zod.object({
  title: zod.string(),
  targetAmount: zod.number(),
  currentAmount: zod.number(),
  deadline: zod.string().nullish(),
  category: zod.string(),
  icon: zod.string(),
  color: zod.string(),
});

/**
 * @summary Update goal progress
 */
export const UpdateGoalParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateGoalBody = zod.object({
  currentAmount: zod.number().optional(),
  title: zod.string().optional(),
  targetAmount: zod.number().optional(),
});

export const UpdateGoalResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  targetAmount: zod.number(),
  currentAmount: zod.number(),
  deadline: zod.string().nullish(),
  category: zod.string(),
  icon: zod.string(),
  color: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Send a message to AI assistant
 */
export const AiChatBody = zod.object({
  message: zod.string(),
  context: zod.string().nullish(),
});

export const AiChatResponse = zod.object({
  reply: zod.string(),
  suggestions: zod.array(zod.string()),
});