# lib/api-zod/src/generated/types/categoryBreakdownItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HelpFinance API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryBreakdownItem {
  category: string;
  amount: number;
  percentage: number;
  color: string;
}