# lib/api-zod/src/generated/types/createGoalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HelpFinance API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGoalBody {
  title: string;
  targetAmount: number;
  currentAmount: number;
  deadline?: string | null;
  category: string;
  icon: string;
  color: string;
}