# lib/api-zod/src/generated/types/createTransactionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HelpFinance API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransactionBodyRecurringInterval } from "./createTransactionBodyRecurringInterval";
import type { CreateTransactionBodyType } from "./createTransactionBodyType";

export interface CreateTransactionBody {
  description: string;
  amount: number;
  type: CreateTransactionBodyType;
  category: string;
  date: string;
  isRecurring: boolean;
  recurringInterval?: CreateTransactionBodyRecurringInterval;
}