# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HelpFinance API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalBalance: number;
  monthlyIncome: number;
  monthlyExpenses: number;
  savingsRate: number;
  topExpenseCategory: string;
  transactionCount: number;
  aiInsight: string;
}