# lib/api-zod/src/generated/types/gamificationProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HelpFinance API
 * OpenAPI spec version: 0.1.0
 */
import type { Badge } from "./badge";

export interface GamificationProfile {
  xp: number;
  level: number;
  levelName: string;
  xpToNextLevel: number;
  streak: number;
  badgeCount: number;
  earnedBadges: Badge[];
}