# lib/api-zod/src/generated/types/updateGoalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HelpFinance API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateGoalBody {
  currentAmount?: number;
  title?: string;
  targetAmount?: number;
}